Deleção de cliente error 404

+++ src/main/java/com/jlucas/client/services/ClientService.java
@@ -1,6 +1,7 @@
 package com.jlucas.client.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -9,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.jlucas.client.dtos.ClientDTO;
 import com.jlucas.client.entities.Client;
 import com.jlucas.client.exceptions.ClientNotFoundException;
+import com.jlucas.client.exceptions.DatabaseException;
 import com.jlucas.client.repositories.ClientRepository;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -57,7 +59,15 @@ public class ClientService {
 	
 	@Transactional
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ClientNotFoundException("Cliente não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de Integridade referencial");
+		}
+		
 	}
